% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csdf-class.R
\name{as.data.frame.csdf}
\alias{as.data.frame.csdf}
\title{S3 method to get data.frame}
\usage{
\method{as.data.frame}{csdf}(x, ...)
}
\arguments{
\item{x}{a csdf object}

\item{...}{ignored}
}
\value{
\code{\link{data.frame}} from \code{\link{slot}} \emph{data}.
}
\description{
S3 method to get data.frame
}
\examples{
Sys.setenv(TZ='GMT')
fpath <- system.file("extdata", "Station_Daily.dat", package="csdf")
obj <- read.toa5(fpath)
summary(as.data.frame(obj))
}

