% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csdf-class.R
\docType{class}
\name{csdf-class}
\alias{csdf}
\alias{csdf-class}
\title{Class encapsulating content of Campbell Scientific data file}
\description{
Objects of this class are typically produced by \code{\link{read.toa5}}
function.
}
\section{Slots}{

\describe{
\item{\code{data}}{\code{\link{data.frame}} with captured data.}

\item{\code{variables}}{\code{\link{data.frame}} with rows units and processing
instruction (Smp, Avg, Min, Max and such).}

\item{\code{meta}}{\code{\link{data.frame}} with auxiliary information.}
}}

