% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.dt.R
\name{read.dt}
\alias{read.dt}
\title{Read Decagon DataTrac 3 SQLite files}
\usage{
read.dt(file, process = TRUE)
}
\arguments{
\item{file}{a dot db file with SQLite data used by DataTrac3. Perhaps you can
find some at \code{\%LocalAppData\%\\Decagon\\DataTrac\\databases\\New Devices\\} .}

\item{process}{ignored}
}
\value{
an S4 class \code{\linkS4class{csdf}}
}
\description{
This is a very limited implementation to read GS3 sensor data out of DataTrac
(SQLite) database files directly. This function uses default calibration
curves and will issue a warning if calibration data are found in a database.
}
\examples{
fpath <- system.file("extdata", "DataTrac3.db", package="csdf")
obj <- read.dt(fpath)
\dontrun{
write.toa5(obj, "somewhere.dat")
}
}

