% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.tob1.R
\name{read.tob1}
\alias{read.tob1}
\title{Read Table Oriented Binary 1 file}
\usage{
read.tob1(file)
}
\arguments{
\item{file}{a file name to read}
}
\value{
an S4 class \code{\linkS4class{csdf}}
}
\description{
Read Table Oriented Binary 1 file
}
\examples{
Sys.setenv(TZ='GMT')
fpath <- system.file("extdata", "TOB1_Station_Daily.dat", package="csdf")
obj <- read.tob1(fpath)
}

