% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.ulong.R
\name{read.ulong}
\alias{read.ulong}
\title{Read unsigned long from connection as numeric}
\usage{
read.ulong(con, endian)
}
\arguments{
\item{con}{a \code{\link{connection}} object. It cannot be a character string naming a
file or a raw vector as it reads by words and will start over for the second word.}

\item{endian}{The endian-ness ("big" or "little") of the target system for
the file. Files produced with LoggerNet on MS Windows are little-endian.}
}
\value{
numeric
}
\description{
Read unsigned long from connection as numeric
}

