% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\docType{methods}
\name{summary,csdf-method}
\alias{summary,csdf-method}
\title{Get common information about stored data}
\usage{
\S4method{summary}{csdf}(object)
}
\arguments{
\item{object}{\code{\linkS4class{csdf}} object to get summary for}
}
\value{
a \code{\link{list}} summarizing stored data with \code{summary.csdf}
  \code{\link{class}} assigned to trigger \code{\link{print.summary.csdf}} S3
  method.

List components are:
\describe{
\item{\emph{granularity}}{\code{\link{difftime}} for smallest positive
non-zero interval between measurements.}
\item{\emph{ts.ok}}{\code{TRUE} if TIMESTAMP column in @data
slot is monotonic. If \code{FALSE}, potentially unintentional clock
correction might have taken place, or other serious data problem.}
\item{\emph{rn.ok}}{\code{TRUE} if RECORD is sequential in slot data of csdf object.
CR program reload results in RECORD counter reset.
If TOA5 file contains restarts it may indicate that
parts of a file were produced with a different CR Basic program.}
}
\item{\emph{nogap}}{\code{TRUE} if adjacent records are \emph{granularity}
apart. You might have issues if \code{FALSE}.}
\item{\emph{gaps}}{a \code{\link{vector} containing TIMESTAMP befor gaps.}}
\item{\emph{from}}{initial/minimal TIMESTAMP}
\item{\emph{to}}{final/maximal TIMESTAMP}
}
\description{
Get common information about stored data
}

