% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.toa5.R
\name{write.toa5}
\alias{write.toa5}
\title{Write TOA5 file from valid csdf class}
\usage{
write.toa5(x, file = "", fileEncoding = "")
}
\arguments{
\item{x}{\code{\link{csdf-class}} object}

\item{file}{a file to write}

\item{fileEncoding}{character string: if non-empty declares the encoding to
be used on a file (not a connection) so the character data can be
re-encoded as they are written. See \code{\link{file}}.}
}
\description{
Write TOA5 file from valid csdf class
}

