% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.csixml.R
\name{read.csixml}
\alias{read.csixml}
\title{Read Campbell Scientific XML}
\usage{
read.csixml(file)
}
\arguments{
\item{file}{The name of the file containing the Campbell Scientific XML
contents. This can contain \~ which is expanded to the user's home
directory.}
}
\value{
an S4 class \code{\linkS4class{csdf}}
}
\description{
Read Campbell Scientific XML
}
\examples{
Sys.setenv(TZ='GMT')
fpath <- system.file("extdata", "CSIXML_Station_Daily.dat", package="csdf")
obj <- read.csixml(fpath)
}

