% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.toa5.R
\name{read.toa5}
\alias{read.toa5}
\title{Read TOA5 file}
\usage{
read.toa5(file)
}
\arguments{
\item{file}{the name of the file to read}
}
\value{
an S4 class \code{\linkS4class{csdf}}
}
\description{
This function reads CSV-alike TOA5 file. It assumes that \emph{TIMESTAMP}
column and those ending with \emph{_TMx} and \emph{_TMn} contain time stamps
in \strong{"YYYY-MM-DD HH:MM:SS"} format.
}
\details{
It relies heavily on \code{\link{as.POSIXct}} for time stamp recognition as
used internally by \code{\link{read.table}}. Therefore it is important to
make sure correct time zone is set so you won't get unexpected conversions.
One approach is to use \code{Sys.setenv(TZ='GMT')} in the code. Use something
like this if you chose to stick to local time zone but decided to neglect DST
switch.

No column name sanitation is performed. As a result, column names are likely
not syntactically valid if arrays were used in CR Basic. To reference such
columns, use backquote marks, e.g., \code{obj@data$`Air(1)`}.

"NAN" is treated as NA.
}
\examples{
Sys.setenv(TZ='GMT')
fpath <- system.file("extdata", "Station_Daily.dat", package="csdf")
obj <- read.toa5(fpath)
}

