% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.toa5.R
\name{write.toa5}
\alias{write.toa5}
\title{Write TOA5 file from valid csdf class}
\usage{
write.toa5(x, file = "")
}
\arguments{
\item{x}{\code{\link{csdf-class}} object}

\item{file}{a file to write}
}
\description{
Write TOA5 file from valid csdf class
}

